Add config

+++ Content/Projectiles/GlobalModifications/SlimeGodMineGlobal.cs
@@ -1,5 +1,6 @@
 ﻿using CalamityMod.NPCs.SlimeGod;
 using CalamityMod.Projectiles.Boss;
+using CataclysmMod.Common.Configs;
 using Terraria;
 using Terraria.ModLoader;
 
@@ -9,7 +10,7 @@ namespace CataclysmMod.Content.Projectiles.GlobalModifications
     {
         public override void PostAI(Projectile projectile)
         {
-            if (projectile.type == ModContent.ProjectileType<AbyssMine>())
+            if (CalamityChangesConfig.Instance.abyssMinesExplode && projectile.type == ModContent.ProjectileType<AbyssMine>())
                 if (!NPC.AnyNPCs(ModContent.NPCType<SlimeGodCore>()) && !NPC.AnyNPCs(ModContent.NPCType<SlimeGod>()) && !NPC.AnyNPCs(ModContent.NPCType<SlimeGodSplit>()) && !NPC.AnyNPCs(ModContent.NPCType<SlimeGodRun>()) && !NPC.AnyNPCs(ModContent.NPCType<SlimeGodRunSplit>()))
                     projectile.Kill();
         }
